+++ packages/preferences-controller/src/PreferencesController.ts
@@ -9,6 +9,7 @@ import type {
   KeyringControllerState,
   KeyringControllerStateChangeEvent,
 } from '@metamask/keyring-controller';
+import type { Hex } from '@metamask/utils';
 
 import { ETHERSCAN_SUPPORTED_CHAIN_IDS } from './constants';
 
@@ -140,6 +141,12 @@ export type PreferencesState = {
    * User to opt in for smart account upgrade for all user accounts.
    */
   smartAccountOptIn: boolean;
+  /**
+   * User to opt in for smart account upgrade for specific accounts.
+   *
+   * @deprecated This preference is deprecated and will be removed in the future.
+   */
+  smartAccountOptInForAccounts: Hex[];
 };
 
 const metadata = {
@@ -164,6 +171,7 @@ const metadata = {
   privacyMode: { persist: true, anonymous: true },
   dismissSmartAccountSuggestionEnabled: { persist: true, anonymous: true },
   smartAccountOptIn: { persist: true, anonymous: true },
+  smartAccountOptInForAccounts: { persist: true, anonymous: true },
 };
 
 const name = 'PreferencesController';
@@ -246,6 +254,7 @@ export function getDefaultPreferencesState(): PreferencesState {
     privacyMode: false,
     dismissSmartAccountSuggestionEnabled: false,
     smartAccountOptIn: true,
+    smartAccountOptInForAccounts: [],
   };
 }
 
@@ -621,6 +630,19 @@ export class PreferencesController extends BaseController<
   setSmartAccountOptIn(smartAccountOptIn: boolean) {
     this.update((state) => {
       state.smartAccountOptIn = smartAccountOptIn;
+    });
+  }
+
+  /**
+   * Add account to list of accounts for which user has optedin
+   * smart account upgrade.
+   *
+   * @param accounts - accounts for which user wants to optin for smart account upgrade
+   * @deprecated This method is deprecated and will be removed in the future.
+   */
+  setSmartAccountOptInForAccounts(accounts: Hex[] = []): void {
+    this.update((state) => {
+      state.smartAccountOptInForAccounts = accounts;
     });
   }
 }